Adds the ability to read the typescript files
This addition is not meant to be full support of typescript from the library. It's meant to allow the reading of the files in the case you are looking to use TS files and are responsible for the transpilation/compilation of those files.

We are experimenting adding this to allow us to use ts files with ts-jest and other tooling in our tests and source files. The compilation is happening locally and doesn't depend on the package having full support for a build pipeline.

This will be published a major version so that those who may have .ts files won't have a break in their dev flow but once it's promoted out of release candidate, it will have proper documentation if we opt to keep the support.