Merges the 2.0.0 release candidate branch (#89)
* Adds the ability to read the typescript files (#76)

This addition is not meant to be full support of typescript from the library. It's meant to allow the reading of the files in the case you are looking to use TS files and are responsible for the transpilation/compilation of those files.

We are experimenting adding this to allow us to use ts files with ts-jest and other tooling in our tests and source files. The compilation is happening locally and doesn't depend on the package having full support for a build pipeline.

This will be published a major version so that those who may have .ts files won't have a break in their dev flow but once it's promoted out of release candidate, it will have proper documentation if we opt to keep the support.

* v2.0.0-rc.0 (#77)

* feat(*): [RO-18670] add ability to ignore dependencies (#78)

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
add support for skipping dependency injection by setting

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
improve logging output when warning against dependency registration collisions

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
fix passing incorrect parameters
improve logging

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
add warning for ignored dependencies

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
add test cases for registerFolders and addResolvableDependency

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
add test cases for addDependency

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
update FileFilterExpression to ignore .d.ts files
add test

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
fix file filter causing issues with some imports
update logging to print primitive types

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration
added missing lodash.get dependency

* feat(*): RO-18670 Extend spur-ioc to add the ability to skip dependency registration (#80)

fix logging

* v2.0.0-rc.1 (#79)

* refactor(warnings): Removes the noisy ignored dependency warnings (#81)

* v2.0.0-rc.2 (#82)

---------

Co-authored-by: Terry Lewis <terry1994dev@gmail.com>