refactor(colors): Removal of chalk coloring messages (#115)
The colorization of these messages do not add anything substantial to the logs unless they are running locally.

Removing it does however helps make this package more secure. There has been a few security concerns that chalk has exposed this framework to because of the low maintenance effort.

While spur-ioc and all other spur-packages will be going into maintenance mode, we are actively reducing any change that a new security concern can be discovered down the road due to a dependency.

If you want to see the colors, you can stick to the previous version of spur-ioc.